refactor: Reimplement heating bill generation API using a structured manager, context, and step-based workflow.
