+++ src/app/api/heating-bill/generate/steps/BaseStep.ts
@@ -1,0 +1,9 @@
+import { HeatingBillContext } from "../context/HeatingBillContext";
+
+export interface IHeatingBillGenerationStep {
+    execute(context: HeatingBillContext): Promise<void>;
+}
+
+export abstract class BaseStep implements IHeatingBillGenerationStep {
+    abstract execute(context: HeatingBillContext): Promise<void>;
+}
